factory: Ensure generated names are unique
This commit ensures the generated name for a type is unique. When adding
sublcass support we neglected to take naming into account resulting in
types for two different base classes selecting the same name.

We add a function, MakeName, which takes the type identity and
returns the target interface's namespace concatenated with the interface
name (minus the first leading I) and the parent class name.