thallium: Move throw to handler pipeline
This commit moves the default failure behaviour into the handler
pipeline, where it should have been originally were it not for this
author's ignorance.

The improvement allows us to do some nice things with Polly so we
provide a static class for constructing retry and circuit break
behaviours on certain exception conditions based on the behaviours of
the new ThrowOnFail handler, which replicates the default behaviour of
RestClient, before its refactoring into BaseClient.