thallium: Further API refinement
The prior commit, 0f4dc64, attempted to refine the interface but didn't
take the correct approach - the introduced generic factories aren't very
accessible because they're not expressed on the runtime API access
point.

This commit restores some old method signatures to avoid an API break,
marks all methods of the API class to do with emission and construction
as Obsolete and moves those, in addition to the V1 introspection
delegate into a separate file - where they can live until we bump the
library's major version.

Finally, we introduce two new interfaces, IApi and IApi't which allow us
to provide an access point on the Api class which won't get out of hand
in terms of method signature. The generic version of the interface
allows us to encode a type parameter for the base class which we can
then use to target extension methods which expose the construction
signatures that we expect.