thallium: Assertion update
This commit adds a factory assertion on the call information coming back
from introspectors allowing us to move the verb check there and remove
it from the attribute inspection logic.

In addition we use the C#6 nameof feature to add some future proofing to
error messages.

Finally we add some default initialisation of sensitive properties on
the call information and move the patch for a null static headers
collection to that region.