thallium: Don't dispose client after each call
This commit swaps the per call creation of an HttpClient (and handler,
in the case of the BaseClient implementation) for a single lazy instance
during a client's life time which also disposes its handler, simplifying
the dynamic and rest subclasses a little bit. This tweak also allows us
to drop the async modifier from the SendAsync method, reducing
complexity a bit further in the base class.

We don't go all out static, to keep some separation for people wishing
to use its features in a specific way.

This should avoid performance degradation [1] through better resource
usage as connections can be shared between calls.

This also involves adding IDisposable to the BaseClient which is safe
to be exposed via your own interface declaration.

[1] http://aspnetmonsters.com/2016/08/2016-08-27-httpclientwrong/