thallium: Use Client Timeout
This commit restructures the defaults configuration and pushes more of
it to deferred initialisation and within BaseClient for ease of
maintenance, returning null in the virtual cover methods.

We add the timeout parameter to the dynamic configuration interface but
this must be set before the first request, as with the handler chain. A
new construction signature is added to RestClient and static
construction extensions exposing the parameter.

Sadly changes to extension method signatures will likely break ABI
compatibility and require a recompile.