thallium: Allow client factories
This commit supports client factories which sort of runs in the face of
our performance concerns.

To get the benefit of both we move the lazy construction function to the
private interface and call the virtual Client method again to get the
client from SendAsync. In the base class this simply referrs to the
lazily initialised field.

In addition we unseal the configuration class subtypes.