thallium: Handlers, new base implementation
This commit refactors the throwing handler into a base/subclass setup
with the base containing an empty set of predicates to throw on and the
old default base class defining throws when the response status is
unsuccessful. If there are no predicates then the new base class will
always throw.

We change the behaviour of the other ResponseHandlers to match this
least surprise idea - not calling On will result in the handler
executing its prescribed action.