thallium: Leading slashes, non breaking tweak
When performed, we added a configuration item allowing reversal of the
breaking behaviour change. We improve on that here and allow the
consumer to set the introspector used in the static Api class when none
is provided.

This should allow mixing new and old behaviour interfaces, although
doing so would still be some work.