thallium: Templating perf, Ignore => Explict
These attributes behave differently in VS - Ignore doesn't run whether a
TestFixture is selected or not from the unit runner window. Suspect this
is something to do with the tests inheriting the ignored state from the
fixture and refusing to run. Anyway, we can run them without source
changes now.