thallium: JSON media type formatter improvements
We wanted to subclass this component and use a custom JSON serializer
with it but that wasn't possible due to bad design.

This commit allows consumers to provide their own JSON serializer and
should yield a performance improvement as we pass the incoming streams
directly to the serializer instead of reading to string values first.