Added Payload response support in SkeletonClient.
This commit adds new type - `Response<,>` that can hold both success and
fail objects, as well as the status code. It can be used in cases where
the consumers of Thallium don't want to parse manually content when
exceptions is catched, or don't want Thallium to throw exceptions at
all.